requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cv==1.0.0
cycler==0.12.1
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-restframework==0.0.1
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.4
fonttools==4.52.4
google-ads==24.0.0
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.0
google-cloud-storage==2.17.0
google-cloud-texttospeech==2.16.3
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
kombu==5.4.0
matplotlib==3.9.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.23.6
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
praat-parselmouth==0.4.3
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
whitenoise==6.6.0
wkhtmltopdf==0.2
